TwitchBot/pyproject.toml
[tool.poetry]
name = "skyrimtwitchbot"
version = "0.1.0"
description = ""
authors = ["Mrowr Purr <mrowr.purr@gmail.com>"]
license = "0BSD"
readme = "README.md"
packages = [{include = "skyrim_twitch_bot", from = "src"}]

[tool.poetry.scripts]
twitch-bot = "skyrim_twitch_bot.__main__:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
twitchio = "^2.8.2"
PySide6 = "^6.6.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.development.dependencies]
poethepoet = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
